rewrote link structure
Rewrote QueryOpts.pm link structure to correctly use qopts of target template 
instead of the qopts for the template we're in.

Got rid of the idea of classes for qopts.

Added first stab at persistant in-mem cache to speed things up since we're 
using more template objects with the new link system.


git-archimport-id: e@ericrichardson.com--2004-eThreads/eThreads--eTrevolution--2.0--patch-8