Moved more data to memcache
* Stuck more data in the memcache
* Added (get|set)_raw routines to memcache for storing non-objects
* changed from SizeAwareMemoryCache to FastMemoryCache:

I like the idea of SAMC better, but the reliance on Storable::dclone is 
killing performance, ie taking 40% of run time.  This'll do for now.

Here's the question now:  How do you have a persistant memcache and make 
sure it stays current with the db when multiple processes are keeping caches? 
I don't have that solved right now.


git-archimport-id: e@ericrichardson.com--2004-eThreads/eThreads--eTrevolution--2.0--patch-9